requirements.txt
numpy
Pillow
tqdm
tabulate
matplotlib
ipython
opencv-python